# Conflicts:
#	java/progettoSweng/.idea/workspace.xml
#	java/progettoSweng/src/com/jsonrpc/ErrorHandler.java
#	java/progettoSweng/src/com/jsonrpc/JsonRpcBatchRequest.java
#	java/progettoSweng/src/com/jsonrpc/JsonRpcDefaultError.java
#	java/progettoSweng/src/com/jsonrpc/JsonRpcManager.java
#	java/progettoSweng/src/com/jsonrpc/JsonRpcResponse.java

+++ java/progettoSweng/src/com/sweng/Node.java
@@ -44,7 +44,7 @@ public class Node {
 
         // Read response from Broker
         JSONObject result = response.getResult();
-        boolean serviceRegistered = (boolean) result.get("serviceRegistered");
+        boolean serviceRegistered = (boolean)result.get("serviceRegistered");
         if (serviceRegistered) {
             String newMethodName = (String) result.get("methodName");
             service.getServiceMetadata().setMethodName(newMethodName);
@@ -80,14 +80,17 @@ public class Node {
     }
 
     public void deleteService(String method) { // missed in uml class diagram
-        IConnection connection = this.connectionFactory.createConnection();
-        JsonRpcManager manager = new JsonRpcManager(connection);
-        manager.sendNotification("deleteService");
-
-        // Delete service
-        RunningService availableService = this.ownServices.get(method);
-        availableService.delete();
-        this.ownServices.remove(method);
+        if (this.ownServices.containsKey(method)) {
+            IConnection connection = this.connectionFactory.createConnection();
+            JsonRpcManager manager = new JsonRpcManager(connection);
+            manager.sendNotification("deleteService");
+            // Delete service
+            RunningService availableService = this.ownServices.get(method);
+            availableService.delete();
+            this.ownServices.remove(method);
+        } else {
+            throw new RuntimeException("There is no service named " + method);
+        }
     }
 
     public void setConncetionFactory(IConnectionFactory connectionFactory) {
@@ -116,7 +119,7 @@ public class Node {
         ArrayList<ServiceMetadata> list = new ArrayList<ServiceMetadata>();
         list.clear();
 
-        JsonRpcResponse response = this.requestService("searchStrategy", searchStrategy.toJson());
+        JsonRpcResponse response = this.requestService("getServicesList", searchStrategy.toJson());
 
         JSONObject json = response.getJsonRpc();
         JSONArray array = (JSONArray) json.get("result");
@@ -127,8 +130,8 @@ public class Node {
         return list;
     }
 
-    private int generateNewId() {
-        return this.id++;
+    private ID generateNewId() {
+        return new ID(this.id++);
     }
 
     // End of Service requester functionality
